Extract dashboard loaders into src/core/dashboard.ts
Move hub/mod dashboard loaders, viewer snapshot, and hub/mod-panel
state payload builders into a dedicated module. Typecheck, 167 tests,
and build all pass.
